Change exam to use assessment service to find accommodations and windows
